Cast link limits to int Close #12
Close #12

+++ javasdk/NRSDK/src/com/neuronrobotics/replicator/driver/BowlerBoardDevice.java
@@ -163,8 +163,8 @@ public class BowlerBoardDevice extends GenericPIDDevice implements ILinkFactoryP
 								conf.getHardwareIndex(),
 								conf.getScale(),
 								conf.getIndexLatch(),
-								conf.getLowerLimit(),
-								conf.getUpperLimit()}, 5);
+								(int)conf.getLowerLimit(),
+								(int)conf.getUpperLimit()}, 5);
 		ConfigurePIDController(conf.getPidConfiguration());
 		
 		return;
